+++ .php-cs-fixer.dist.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 $finder = PhpCsFixer\Finder::create()
     ->in([
@@ -7,11 +8,12 @@ $finder = PhpCsFixer\Finder::create()
     ])
     ->name('*.php');
 
-return (new PhpCsFixer\Config())
+return new PhpCsFixer\Config()
     ->setRiskyAllowed(true)
     ->setRules([
         '@PSR12' => true,
         '@PHP8x4Migration' => true,
+        '@PHP8x2Migration:risky' => true,
         'array_syntax' => ['syntax' => 'short'],
         'declare_strict_types' => true,
     ])